SS4 compat: Rename code to src, tabs to spaces, implement namespaces, PSR-4 autoloader
