+++ .upgrade.yml
@@ -1,0 +1,5 @@
+mappings:
+  ApiKeyMemberExtension: Sminnee\ApiKey\ApiKeyMemberExtension
+  ApiKeyRequestFilter: Sminnee\ApiKey\ApiKeyRequestFilter
+  GridFieldAddApiKeyButton: Sminnee\ApiKey\GridFieldAddApiKeyButton
+  MemberApiKey: Sminnee\ApiKey\MemberApiKey
